requirements.txt
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
docutils
pygments
pypiwin32
kivy
selenium



